[IOTMID-154] Adds timestamps, filters to endpoints
This commit adds new created/updated fields as to
allow the user to keep track of changes to the
configuration of a device/template.

Also, this adds new query string parameters to the
GET /device and GET /template endpoints, allowing
the user to either sort or limit the resulting set.

The query string parameters are:
- limit=<#> : limits the number of entries in the response
- sortAsc=<field> : returns the set in ascending order of the
                    given field
- sortDsc=<field> : returns the set in descending order of the
                    given field

Change-Id: Ic98e2cfb78a1fc91664a3afff04ab78f8baed83f