[IOTMID-287] Implements helper endpoints for semi-smart devices
This commit adds a simple endpoint to device manager
allowing the caller to find a device given a set of
constraints.

This should help the implementation of a new flow on
automate that determines the MQTT topic to be used
based solely on the "human readable" label of a device.

Change-Id: I538441392214e647b93f8d55bca5e63f2053db2e