[IOTMID-293] Adds device status from ctx broker to device data
This commit makes the devices' data returned from GET /devices
include status information, collected from the context broker,
allowing applications to skip the explicit call to the broker.

While functional, this is actually a sign of bad responsability
distribution amongst services - a service should not need to
make explicit calls to other services in order to fulfill a
client's request. This leads to tricky dependency management and
increases service coupling throughout the solution.

Change-Id: I48011d94bcca22e5a27c3f438f3331268c45a945