Renames module, improves initialization sequence
This renames the module, thus making linter happier. Also, and
all the more important, this improves the startup script used in
the container allowing it to better check the liveliness of
psql on service boot, as well as allowing it to automatically
create the required database if it doesn't exist already.

This fixes dojot/dojot#99.