Fixes bogus trailing counter, automatic template on delete cascade
This commit fixes two issues found during the tests of 0.2.0-nightly20180119
build.

First, it removes the bogus counter being added when a single
device creation request is performed.

Also, this changes relation configuration between devices and
templates to prevent the automatic removal of templates from
devices, when a template is removed.