Adds missing commits from schema creation/changes
This change adds a couple of extra commits to the tenant initialization
procedure, for safekeeping.

When a tenant's first call to device manager is a READ command (e.g
"GET /device"), the tenant's context is created (`db.create_all()`)
but the trigger configuration is not properly committed to the db,
as there's no other data to commit.

Since the tenant's context is already created a WRITE command (e.g.
"POST /device") is run using the previously created namespace schema,
which, at this point, does not contain the required triggers to
validate inter-template attribute uniqueness.