Fixing Kafka group ID for status monitor
Sometimes, if device manager crashes (or is restarted), or even if more
than one tenant is created, all status monitors consumers will use the
same group ID. Kafka will then send messages to one consumer of this
group, which is not what was intendend. This commit creates a different
group ID each time a new consumer is created - which includes status
monitors for different tenants.